notes feature, thought ids still not connected to philosopher

+++ src/services/ApiService.js
@@ -8,6 +8,8 @@ class ApiService {
 
   getPhilosophers= () => fetch(this.api + "/philosophers").then(res => res.json())
   
+  getThoughts= () => fetch(this.api + "/thoughts").then(res => res.json())
+  
   createPhilosopher = (newPhilosopher) => {
     newPhilosopher.user_id = user.id
     return fetch(this.api + "/philosophers", {
@@ -20,6 +22,19 @@ class ApiService {
     .then(response => response.json())
   }
 
+  createThought = (newThought) => {
+    newThought.user_id = user.id
+    return fetch(this.api + "/thoughts", {
+      method: 'POST', // or 'PUT'
+      headers: {
+        'Content-Type': 'application/json',
+      }, 
+      body: JSON.stringify(newThought),
+    })
+    .then(response => response.json())
+  }
+  
+  
   findOrCreateUser = (username) => {
     return fetch(this.api + "/users", {
       method: 'POST', // or 'PUT'
@@ -35,19 +50,28 @@ class ApiService {
    
   
     deletePhilosopherAction = (id) => {
-      
-      return fetch(`${this.api}/philosophers/${id}`,{
-    
-      method: "DELETE",
-      }).then(() => {
-        console.log('removed')
-      }).catch(err => {
-        console.error(err)
-      })
-      
-     
+      let result = confirm("Do you want to delete this movie?")
+      if (result){
+      let configObj = {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+          "Accept": "application/json"
+        },
+        body: JSON.stringify({
+          id: this.id
+        })
+      };
+      fetch(`${this.api}/philosophers/${id}`, configObj)
+
+.then(resp=>resp.text())
+.then(resp=>alert(resp))
+.then(target.parentNode.removeChild(target))
+.catch(error=>console.log(error))
+}    
+}
       
   
-  }
+
 }
